refactor : 전적 조회 API를 MemberRecord Layer로 이동
